node-main/deps/v8/test/webkit/fast/js/Promise-resolve-chain-expected.txt

Test chained Promise resolutions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "hello"
PASS result is "bye"
PASS successfullyParsed is true

TEST COMPLETE
